split views into views and tool_views
