+++ templates/base.html
@@ -26,7 +26,7 @@
 <body>
 
   <nav class="navbar navbar-expand-lg navbar-dark bg-dark ">
-          <a class="navbar-brand" href='{% url "mainapp_URLs:home" %}'>Home</a>
+          <a class="navbar-brand" href='{% url "home" %}'>Home</a>
           <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExample09" aria-controls="navbarsExample09" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
           </button>
