+++ templates/user/profile.html
@@ -9,7 +9,8 @@
 
            <div class="modal-body text-center">
 
-               <img src="{{user_profile.profilepic.url}}" name="aboutme" width="140" height="140" border="0" class="img-circle"></a>
+               <img src="{{user_profile.profilepic.url}}" width="140" height="140" border="0" class="img-circle">
+
                <h3 class="media-heading">{{user_profile.user.username}}<small>{{user_profile.name}}</small></h3>
                <span><strong>Links: </strong></span>
                 <a href="{{user_profile.github}}" target="_blank"><span class="label label-info">Github</span></a>
